lavc/v4l2-common: Add pack_flags to v4l2 conversion tables and extend the mapping function to use it.
It serves to distinguish if we will (de)packetize the format from an AVFrame or an AVPacket. We will need AVFrame support for v4l m2m and we can certainly not use as-is e.g. V4L2_PIX_FMT_YUV420 as AV_PIX_FMT_YUV420P since the former is pseudo-packed while the later is planar. Also, V4L multi planar formats cannot be packetized into raw video AVPackets.